# docs/openapi.json
{
  "openapi" : "3.1.0",
...
  "components" : {
    "schemas" : {
...
      "CreateTrainingRequest" : {
        "type" : "object",
        "required" : [ "teamId", "trainingDate" ],
        "properties" : {
          "teamId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "trainingDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "startTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "endTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "location" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/TrainingStatus"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
      "LocalTime" : {
        "type" : "string",
        "format" : "local-time",
        "examples" : [ "13:45:30.123456789" ],
        "externalDocs" : {
          "description" : "As defined by 'partial-time' in RFC3339",
          "url" : "https://www.rfc-editor.org/rfc/rfc3339.html#section-5.6"
        }
      },
...
      "TeamDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "seasonId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "seasonName" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "TrainingDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "team" : {
            "$ref" : "#/components/schemas/TeamDto"
          },
          "trainingDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "startTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "endTime" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "location" : {
            "type" : "string"
          },
          "notes" : {
            "type" : "string"
          },
          "status" : {
            "$ref" : "#/components/schemas/TrainingStatus"
          },
          "createdAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "updatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          }
        }
      },
...
      "TrainingStatus" : {
        "type" : "string",
        "enum" : [ "COMPLETED", "CANCELLED" ]
      },
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    }
  },
  "paths" : {
...
    "/api/v1/trainings" : {
      "get" : {
        "summary" : "Find trainings",
        "description" : "Returns trainings matching the optional search criteria.",
        "tags" : [ "Trainings" ],
        "parameters" : [ {
          "name" : "dateFrom",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/LocalDate"
          }
        }, {
          "name" : "dateTo",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/LocalDate"
          }
        }, {
          "name" : "status",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/TrainingStatus"
          }
        }, {
          "name" : "teamId",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "trainingDate",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/LocalDate"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Trainings found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TrainingDto"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "Create training",
        "description" : "Creates a training for a team.",
        "tags" : [ "Trainings" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTrainingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Training created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingDto"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Team not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/trainings/{id}" : {
      "get" : {
        "summary" : "Find training by id",
        "description" : "Returns a single training by id.",
        "tags" : [ "Trainings" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Training found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TrainingDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Training not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "Delete training",
        "description" : "Deletes a mistakenly created training including its participants.",
        "tags" : [ "Trainings" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Training deleted"
          },
          "404" : {
            "description" : "Training not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "info" : {
    "title" : "Coach Buddy API",
    "version" : "1.0.0",
    "description" : "API für Trainings-, Team- und Personenverwaltung"
  }
}